Added a delete statement to the SQL set:
This commit includes:
* lexer changes
* parser changes
* a test suite

The SQL statement is fully validated by the parsers so there are no error
conditions that are returned at the lexer/parser stage.

Validation that the WHERE clause covers the key and the table exists will be
done in the query sub-system at run time, obviously.

In addition a new function has been added to the DDL helper functions
and a new DDL helper utility module has been added to make extending that
easier.

Some minor clean up was done:
* we used to use the same local/partition key in some tests for convenience
  these fail validation now and have been removed