Rework minimum capabilities
There's a capability gap (ha!) in blob: if a table is defined with a
blob type while the cluster is in a mixed mode with a 1.4 node, the 1.4
node when upgraded to 1.5 will not know that the type is "really" a blob
type rather than varchar.

We have plans to make the DDL handling smarter, but for now enforce a
DDL v2 lower bound on defining tables with blobs.

While improving the minimum capabilities decision-making I discovered
some redundant functionality, so collapse the ddl minimum version
functions.